Добавил BasketService и изменил возвращаемые типы в сервисе Car

+++ Automarket.Service/Interfaces/ICarService.cs
@@ -12,14 +12,14 @@ namespace Automarket.Service.Interfaces
         
         IBaseResponse<List<Car>> GetCars();
         
-        Task<IBaseResponse<CarViewModel>> GetCar(int id);
+        Task<IBaseResponse<CarViewModel>> GetCar(long id);
 
-        Task<BaseResponse<Dictionary<int, string>>> GetCar(string term);
+        Task<BaseResponse<Dictionary<long, string>>> GetCar(string term);
 
         Task<IBaseResponse<Car>> Create(CarViewModel car, byte[] imageData);
 
-        Task<IBaseResponse<bool>> DeleteCar(int id);
+        Task<IBaseResponse<bool>> DeleteCar(long id);
 
-        Task<IBaseResponse<Car>> Edit(int id, CarViewModel model);
+        Task<IBaseResponse<Car>> Edit(long id, CarViewModel model);
     }   
 }
